Add more ask and rollback function

+++ Harden/Windows/src/Crapware-RemoveUseLessSoftware.ps1
@@ -1,17 +1,21 @@
 Get-AppxPackage -Name king.com.CandyCrushSaga
 Get-AppxPackage *3dbuilder* | Remove-AppxPackage
 Get-AppxPackage *officehub* | Remove-AppxPackage
-Get-AppxPackage *skypeapp* | Remove-AppxPackage
 Get-AppxPackage *getstarted* | Remove-AppxPackage
 Get-AppxPackage *zunemusic* | Remove-AppxPackage
 Get-AppxPackage *bingfinance* | Remove-AppxPackage
 Get-AppxPackage *zunevideo* | Remove-AppxPackage
-Get-AppxPackage *onenote* | Remove-AppxPackage
 Get-AppxPackage *people* | Remove-AppxPackage -ErrorAction SilentlyContinue
 Get-AppxPackage *windowsphone* | Remove-AppxPackage
 Get-AppxPackage *bingsports* | Remove-AppxPackage
 Get-AppxPackage *xboxapp* | Remove-AppxPackage
 
+if( (ask "Uninstall OneNote ?" "Uninstall-OneNote.ask") -eq $true ){
+	Get-AppxPackage *onenote* | Remove-AppxPackage
+}
+if( (ask "Uninstall Skype ?" "Uninstall-Skype.ask") -eq $true ){
+	Get-AppxPackage *skypeapp* | Remove-AppxPackage
+}
 reg add HKEY_LOCAL_MACHINE\SOFTWARE\Policies\Microsoft\MicrosoftEdge\Main /v AllowPrelaunch /d 0 /t REG_DWORD /f
 
 # List: Get-AppxPackage
